\name{hdtsdemopkg-package}
\alias{hdtsdemopkg-package}
\alias{hdtsdemopkg}
\docType{package}
\title{\packageTitle{hdtsdemopkg}}
\description{\packageDescription{hdtsdemopkg}}
\section{Package Content}{\packageIndices{hdtsdemopkg}}
\author{\packageAuthor{hdtsdemopkg}}
\section{Maintainer}{\packageMaintainer{hdtsdemopkg}}
\keyword{package}
